Updated the Forms model
The model is now extended from the Hazaar\Model\Strict class and so the fields section is 100% compatible with the field definitions available with Strict Models.  Feel the powar!!!

Also, more stuff is handled in the model itself, along with loading the file and checking it is correct and contains all the required pieces.