Dynamic select inputs and loader tweaks
This commit has a couple of major updates.  We now have dynamically loaded select inputs.  This allows us to load the data in a select from an API call.  This is channelled through the forms model to ensure consistency.

There is also an updated load process to show the loader when actually loading something.  This means that the loader is shown whenever there is one or more AJAX calls in progress, then it is hidden when they are all completed.  This works nicely because AJAX calls for loading the above select items triggers the loader to be displayed so we get a nice ajav loader between page navs (if needed) so the user is no staring at a blank box.