Tweaked the load procedure even more for moar betta!
Nav now increments a loading and calls ready so that the loader DIV stays around on an initial page load and hides only when everything is loaded.  The problem was without this the page would either stop loading too soon or would never stop loading depending on whether the page had dynamic content being loaded.