Added a nice datepicker
If the "format" attribute is used, then the field will become a plain text input but uses the bootstrap-datepicker plugin, along with the input mask plugin to make a fairly nice date input.

If the format attribute is not used, then the input will be a plain old HTML5 date input.