Tweaks to input validation control flow
Inputs can now be validated in two methods.  Synchronously and asynchronously.  

Async allows API calls to the backend to confirm values.  However these API calls can only be done on input change.  This result is then recorded in the input field and used later directly when the synchronouse validation is called.