Completed lookup input support
This turned out to be quite intricate and prompted the changes to hazaar-mvc at release-2.3.19 which includes a new jquery helper class in the dataBinder called dataBinderValue.  This allows us to now store values with labels and depending on how it is used, the value will be used or the label.  For example, INPUT elements will get the value and other non-input elements, such as a DIV, will get the label.  This can of course be overridden.