Added field tags
These tags allow fields to be completely hidden if the tag is not set.  These fields will still be available in the load/save methods but will NEVER be sent to the frontend, unless the tag is set with `$this->setTags()` which must be called in the controller init() method.