Major updates to validation
Everything now runs off of callbacks.  This allows us to do asynchonous validation all the time. Previously we had to switch between async and just skipping the async because when saving we had to wait for the validation result.  Using callbacks all the time it now doesn't matter if we do an async API call or not as the validation callbacks are called at the end of either.  We use a validation queue to make sure that the callback is only called once all fields are validated.  This allows us to mix in both sync and async validations at the same time.