Added default placeholder for select lists
Selects will now always have a default option for null values, ie: a placeholder.  I have also removed the code that disables the default option of the field is required.  We should allow the value to be set back to null if the user wants, but the required field validation will still be enforced.