Fixed default value on lazy load options
Turned out to be quite simple.  We would set null on unpopulated fields.  This is in case something else depends on the lazy load field we need it to be null so that andy 'show' conditions will propagate.  If there is a default though we now use that instead which is a newly defined caveat.