Added required field label indicators
This is pretty cool.  I'm using a style to do this so I shouldn't need to add this to any specific field types.  It should work with all of them that have a label (which should be all).  

I even made it dynamic pretty easily so that if a fields requirements change on the same page a a field that it requires, then the class is toggled.