Compound option label keys
If the options.label element contains a '{{' in it somewhere then it is assumed to be a compound label which means it will be parsed with _match_replace().  This allows labels to be a string containing multiple returned fields instead of just one allowing labels to be concatenated programatically in the frontend instead of requiring this to be done on the server.