Protected fields: Backend functionality
The backend PHP code has been updated to handle protected fields.  I have done this first so that it is easy to test by simply changing a protected field value in the test form and saving it.  I can confirm that the value is not saved.  Only non-protected fields are saved.