Select list data handling updates
Completely rewrote how the ajax data is handled.  Previously we would convert arrays of objects into objects with key/value pairs.  Now we do the opposite and convert objects into arrays of objects with properties "value" and "label" which are the default value key and label key.

This allows us to do some fancy sorting and the frontend filter now works.

I may need to update how this works to specify if the sorting is done by value or by label or by other.