Added support for field definition groups
Surprisingly, alot of what was required was already completed.  I already had field groups processing the 'fields' attribute of an object.  This was actually created dynamically when using arrays of arrays for form field layouts.  All I had to do was get the input data-bind names to generate correctly as layout was already working.

This update required changes to MVC so those will be included shortly.