Added better type conversion for integers
I had a problem with saving and reloading some select input values.  If the select data is an object it get's converted to an array automatically.  Problem is with integer keys they get converted to string by the for() call.  Also, sometimes they are being sent as objects because they keys don't start at zero.

So now, we convert them to arrays of objects always for consistency, and the value is typed if it is an integer so that it is stored as an in on the frontend (previously we left it up to the backend to convert type).  This allows our data load indexOf check to work correctly.