Added data sort and convert methods
The convert method converts a key/value object into an array of objects that can be used to describe data, just as the input_select methods did.  input_select is now using this method perfectly.  Sort is being called in the multi_select method but we need to convert the data.  While this is now trivial, we need to update the input builder code to support the new data format.