Performance update
I did some testing and found that form data was being loaded on all interact commands, sometimes twice (like in 'post') and this was totally unnecessary.  It was because we were calling the form() method, which was meant for the child class to call from it's displaying action because it also sets up helpers, view stuff, etc.  Interact is now just doing the form Model creation and check so things should be a bit faster.