Validate via URL now only if value not null
If the value is null, the URL was being evaluated as missing the __input__ parameter, so the URL would be false causing the validation to throw an error.  Now we not only force the match_replace to return but we only do it if the item has a value.  Technically setting required:true should handle instances where you want to validate the field has a value.