Added custom data objects
It's now possible to specify a custom data object and re-use it as a field.  See #133 for details on usage as I have stuck to the spec in there.

Only thing not yet done is defining custom layouts in the fields.  This will also affect existing field groups so I am committing it separately.