Tags are now usable in Interact
While the fields are still blacked out on the sever, this allows us to do some other stuff in the form layout based on the tags.  Tags are now usable in _eval() so they can be used in things like "show" or "allow_remove" and such.  So I have also added allow_add, allow_remove and allow_edit in list fields as eval options.