Enhanced the attachment()
Calling file_get() now only happens if the method exists.  The reason for this is to trick PHP into allowing any function arguments to be used.  This is because it is up to the implemented file_list() method (be it the built-in one, or user implemented one) to dictate what the target URL looks like.  So haveing a variadic argument list allows this to work better.