Enhanced field layouts
Layouts now operate more consistently and in line with what we currently support for page layout definitions.  This means you can now specify either a string with the field name, an object which may contain either a "fields" or "name" attribute, or an array of the above including arrays of arrays.