Added "keep" option to field defs.
Normally, when a field is being hidden due to a "show" it's value is set the null.  This is so that any fields that reference it in their "show" rules will also be hidden.  However, there are situations where we don't want this.  The "keep" options will cause the value to be kept.  Use this cautiously.