Added field disable/enable when no options
This affects both multi and single selects.  In the process of implementing this I also found some duplicate code for populating single select lists with options.  This is now done using a common items function for consistency and this is the function that does the enable/disable and all that jazz.