Fixed possible recursive loop on toggle show
I had a situation where there were rules on a sub field that triggered an update when nullifying the field, then the update would do something that would trigger the toggle show function which would update the field and it would get caught in a loop.  So now, once we enter the toggle show we set a working state, then before we call it we check if we're already working.  This will prevent entering the loop in the first place.