Fixed possible error on page nav/reload
If you nav away from a page while SELECTs are still loading from an API, then the SELECT will be destroyed on the DOM, but still exist in memory due to references, so when the AJAX call completed some data will not exist.  So now we complete quietly if this happens instead of continuing on to an error.