Enhanced the change event
This was already firing, but only because of jQuery event propagation, which was not goo enough.  Now we stop propagation and then trigger the change event on the form itself and pass the item data that was changed which contains it's value, label, other value and name as well as links to it's parent and stuff.