Modified the required evaluation method
Now, instead of setting up the exclamation when the field input is created then just toggling a class, instead we add/remove the exclamation at evaluation time.  This is because finding subfield of a required field would be problematic on creation, but we should be OK to find them at evaluation time.