Fixed validation on array fields
If the validation result was TRUE but there were no child items, the validation would never return.  It is totally possible for the validation rule to succeed but not have any items, for example when the rules depends on another field, so I added a condition to only check child items if there are any.