Minor tweaks to allow getting a rendered form PDF
Internally this is not used and we keep using the HTML content with a PDF response.  But I shifted some things around and added a new render method to the PDF output class that returns a Hazaar\File object with the rendered content that the user can do whatever it wants with.