Added "local" date format support
Date fields can now specify "local" as the date format.  This will trigger the interface to detect the local date format and use that.  The browsers date format must be set correctly.  

NOTE:  Chrome appears to make it difficult to set this and by default will use US date formats.  Chrome does not adhere to the system date format settings.