Getting fancy.  I now have a node.js code execution server
This server will get started automatically, same as running node.js from the command line, then the forms model will shoot code at it to evaluate.  The advantage here is that once the server is running, the execution time is VERY fast.  This is because it was the node.js startup time that was killing the performance when running it for each evaluation.