More script eval enhancements
Subfields and types all now have an associated value, which also means they now have a parent item.  This means that 'value' and 'item' can be referenced in the object layout.  

NOTE: Logically speaking, processing the layout itself does not descend INTO any fields, so if the script is defined in the layout, then 'value' refers to the type object itself, not 'item' as was previously thought.  Item would then of course be the parent of the type, but given that this relationship would be variable, there's no logical reason to use it, but hey, it exists.