Updated form element item data references
Now, instead of storing the item data itself, we simply store the name of the item.  This was because of a bug caused by copying objects.  The elements that referenced the old item would not be updated to point to the new object item and so events and such would reference old data.  Storing just the name allows us to lookup the data item, get the current one, and use that for events.