Added custom field disabled functions
The disabled eval now uses a stored function to make changes to the form when the disabled value requires evaluation.  The default and previous method still works so this is 100% backwards compatible.  It does this by automatically creating a callback function that implements the previous behavior.  The benefit here is that this function can be overridden, which is what I have started doing here for the list field.