Form list field new item events working
Turned out to be quite a good update.  We just create a new "host" called "sub_host" and start passing that through to the field functions, which means it populates into all the subsequent function calls effectively enabling a "mini form".  Because of this, to reset the fields all we need to do is empty the sub_host.  

I have a feel that this may cause some breakage if there is a top-level field with the same name as a subfield in the list.  The fix for this will be dataBinder namespaces.