Store required result in the form field def
This will be kept up to date when the field changes/updates it's value.  The visual side was working fine so I added a bit to store the required result and then switched over the validation function to use that.  This means that recursive evaluations now work. I think I will need disabled to do the same, possibly.