Fixed field watch inconsistent array element
This is only a problem since we are now no longer cloning the field definitions. Previously the SELECTs would use the watch attribute as a string, then later it would check if the watch attribute was not an array and turn it into one.  With a cloned def, this change disappears almost immediately.  

Now, it persists, so loading a page with a field that has a watch attribute will fail because the SELECT attempts to use it as a string when it is now an array.

To fix, all usage was converted to handle the array, as it should, because that was the original design and was in fact a potential bug if anyone tried to watch multiple fields.  The prepare field def method was then updated to convert the watch attribute to an array if necessary.