Fixed labels in view mode for fields that previously deleted labels
Some fields types, such as checkboxes, use the label for the checkbox, so they do not need a general label.  Previously, we would just delete the label, which was probably quite wrong.  Now we set a `nolabel` flag on the def to let the form_field function know that a label isn't needed. 

Also, in the case of a checkbox, there is a title attribute that can be used.  if this is set, then the title is used in the form_field label instead.