Fixed list item validations
This was because list item fields all share a field definition, however now that we are not cloning them, they are not finding their data correctly.  Validation had a bug where if the data was not found, it would simply return instead of exec the callbacks.  This has also been patched.  But the main fix is when the validation runs the form field defs are cloned before the name is appended, which means the name exists, so the data is located.  Also the validation of sub-fields now uses the fullname and the real host instead of creating a fake one and trying to use the child name.