Added input container to select
Appears all the problems with OTHER inputs on selects are because the SELECT must have a parent and it will not have one if the OTHER input is being shown on a page load.  If the OTHER input is being shown upon change, ie: when the user selects it, there is a parent and everything works.  When the page is loading and a new SELECT is being generated, it is not added to the form yet when the change event is triggered and so has no parent.  Adding the SELECT to a container DIV appears to have fixed this.