Added submit event
This is a special event.  It is triggered BEFORE a save if there are any event handlers.  If there are none, then the save occurs as normal.  If there are event handlers, then it needs to return true for the normal save to occur.  Returning anything (or nothing) other than true will prevent the normal save from running.