Adding dynamic pages
This makes use of most of the current triggers, but adds a new 'pages' trigger for then the pages change.  'nav' is also called if there are any changes as technically the nav position can change.  This allows application implemented page navigation to be updated/redrawn if page structure changes.  If there are no 'show' rules on the pages, then 'pages' will only fire once at the form setup stage.