Added auto option value resolution on toFormArray
Now, if a value is detected that SHOULD have DBV including label, etc, but is just a plain value, the options will be used/looked up to find the label of the value.  This used to happen in the form as per normal use, but if the form defines a static value, this would never automatically have a label and would have to be defined in the form data.  Now, even static values will find their labels automatically. 

Once these lookups are done the result becomes the form data and is saved as such, meaning the lookups should only happen once, or after only when the value changes.