Fixed item_data.update being triggered twice on SELECT
SELECTs are special in that when they are changed, they set the value, then that triggers an update so that the label can be set.  This is done because if the value is set by something else, we still need to lookup and set the label.  Problem is, each time we do the set, we we trigger an update, so on the label set we don't allow an update to be triggered.