Added options data with matchReplace
This allows data to be sent to the options API endpoint without including it in the URL.  This works with both GET and POST and is intended to help send larger amounts of data by allowing POST to send data in the request body.  Previously data could only be sent as part of the URL queryString.  Each data element can be a string with {{mustache}} tags for replacing with form data values.