Added a better fix for #242
By reducing the data down to key/value pairs, we lost any extra data.  This normally wasn't used, but now that we're looking at grouping things, we need the whole data item, so this had to be rethought.  Now, we convert_data earlier, and as part of that build a key/value index that is used to remove item values that don't exist.  Much better as this fixes the original problem, but still gives as access to the full data item.