I have file attachments working in standalone mode
I've decided to include file data as URL encoded data as the 'url' element of a file.  It can then be up to the developer to implement a filter that extracts the URL, decodes it and stores the file, then when form data is loaded again a media URL can be used instead.  Otherwise, the data can be left alone and the encoded data URL will make things work as normal, just maybe a bit slower, due to the size of the form data that now needs to be processed.