Fixed radios loaded from API not updating selected value
Didn't realise that the data resync is actually what set the selected radio.  Thing is, the resync happens in the main thread while the radio options are loaded in async, meaning the options aren't there when the resync happens.  Now we get the item data and set the selected option on the radio while it's being loaded.