Updated _nullify() for code efficiency and fixes keep/protected
Turns out _nullify had duplicate code, which cause the keep/protected check to get overlooked in some cases.  Now instead of looping through sub-fields and dealing with them directly, we offload them back to _nullify to handle recursively allowing more consistent use of code with better functionality.  We now also lookup the def if the one given is a string.

This still needs to be tested on nested objects, which I think will be the next fix.