Fixed select setting label on list item reset breaking "Other"
So it was the reset.  When the select was reset it grabbed the label and set it into the sub_host data item.  Normally this isn't a problem when "other" is not used, because the label is updated again to the correct value.  With other though, this needs to be null.  Simple fix was to check for null on the select value when setting the label.