Removed the layout resolution from Model::resolve
This is more of a display thing and isn't used by all output renderers.  So we don't bother with it here.  The Model resolver should be just about resolving the pages/sections/fields and field values, that's all.  Layouts should be resolved by the renderer which is why this function is moving to the base output class.