Multi-select other values
Initial commit of this so I don't lose it as it's actualy kinda working about 50%.

So far I can show an "Other" option that when clicked, shows a multitext input that can have values added to it.  These values then merge when the save() function is called on the dataBinderArray.  Requires changes to dataBinderArray to support this, but it is working as is.

Need to test/write code for loading data in.  Also, the multitext isn't displaying options correctly, but still seems to function.